pyproject.toml
[tool.poetry]
name = "endpoints"
version = "0.1.0"
description = "A package containing all my LangChain runnables"
authors = ["Mark Hagemann <mark.hagemann@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
openai = "<2"
langchain-community = "^0.2.6"
langchain-core = "^0.2.11"
langchain-openai = "^0.1.20"
langchain-chroma = "^0.1.2"


[tool.poetry.group.dev.dependencies]
langchain-cli = "^0.0.21"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
